+++ 404.php
@@ -2,7 +2,7 @@
 <?php
 /**
  * @link https://www.boxmoe.com
- * @package lolimeow
+ * @package fuwari
  */
 //boxmoe.com===安全设置=阻止直接访问主题文件
 if(!defined('ABSPATH')){echo'Look your sister';exit;}