+++ core/widgets/widget-archive.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @link https://www.boxmoe.com
- * @package lolimeow
+ * @package fuwari
  */
 //=======安全设置，阻止直接访问主题文件=======
 if (!defined('ABSPATH')) {echo'Look your sister';exit;}