+++ page/p-signin.php
@@ -2,7 +2,7 @@
 /**
  * Template Name: 登录页面
  * @link https://www.boxmoe.com
- * @package lolimeow
+ * @package fuwari
  */
 //boxmoe.com===安全设置=阻止直接访问主题文件
 if(!defined('ABSPATH')){echo'Look your sister';exit;}